[Add, Fix] Faradayの導入、APIのサービスクラスリファクタリング #201
