+++ app/services/deepl_api_service.rb
@@ -1,0 +1,29 @@
+class DeeplApiService
+  def initialize(text)
+    @text = text
+  end
+
+  def call
+    translating_japanese_to_english
+  end
+
+  private
+
+  attr_reader :text
+
+  def translating_japanese_to_english
+    connection = Faraday.new('https://api-free.deepl.com') do |builder|
+      builder.request  :url_encoded
+      builder.response :json, :content_type => /\bjson$/
+      builder.adapter  :net_http
+    end
+    params = {
+      text: text,
+      auth_key: Rails.application.credentials.deepl[:api_key],
+      target_lang: 'EN',
+      source_lang: 'JA'
+    }
+    deepl_response = connection.post('/v2/translate', params)
+    translated_text = deepl_response.body['translations'][0]['text']
+  end
+end
